+++ modules/light-clients/06-solomachine/types/header.go
@@ -10,7 +10,7 @@ import (
 	"github.com/cosmos/ibc-go/v3/modules/core/exported"
 )
 
-var _ exported.Header = &Header{}
+var _ exported.ClientMessage = &Header{}
 
 // ClientType defines that the Header is a Solo Machine.
 func (Header) ClientType() string {