+++ modules/light-clients/07-tendermint/types/misbehaviour_test.go
@@ -24,7 +24,6 @@ func (suite *TendermintTestSuite) TestMisbehaviour() {
 	}
 
 	suite.Require().Equal(exported.Tendermint, misbehaviour.ClientType())
-	suite.Require().Equal(clientID, misbehaviour.GetClientID())
 }
 
 func (suite *TendermintTestSuite) TestMisbehaviourValidateBasic() {